requirements.txt
pinecone-client
tiktoken 
openai 
langchain
python-dotenv
InstructorEmbedding
torch
sentence_transformers
fastapi 
uvicorn
pyautogen